% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectNames.R
\encoding{UTF-8}
\name{CollectNames}
\alias{CollectNames}
\title{Collect names an entire list}
\usage{
CollectNames(l, prefix = NULL)
}
\arguments{
\item{l}{A list.}

\item{prefix}{A prefix to nomes.}
}
\value{
The names.
}
\description{
Internal function to collect names an entire list.
}
\keyword{SYNCSA}
