% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flona.R
\docType{data}
\name{flona}
\alias{flona}
\title{Data about flona}
\format{is a list of four matrices}
\source{
Pillar and or Debastiani
}
\usage{
flona
}
\description{
A dataset about plant community composition from Flona
}
\keyword{datasets}

