% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.syncsa.R
\encoding{UTF-8}
\name{organize.syncsa}
\alias{organize.syncsa}
\title{Function for organize data for Package SYNCSA}
\usage{
organize.syncsa(comm, traits, dist.spp, envir, strata)
}
\arguments{
\item{comm}{Community data, with species as columns and sampling units as
rows.}

\item{traits}{Matrix data of species described by traits, with traits as
columns and species as rows.}

\item{dist.spp}{Matrix containing phylogenetic distance between species.
Must be a complete matrix (not a half diagonal matrix).This matrix can be
larger than community data (more species) as long as it has at least all
species that are in community data.}

\item{envir}{Environmental variables for each community, with variables as
columns and sampling units as rows.}

\item{strata}{Strata nomed vector to specify restricting permutations within
species groups.}
}
\value{
The dataframes or matrices of community, traits, phylogenetic distance and
environmental variables. The strata vector for permutations.
}
\description{
Package \strong{SYNCSA} requires that the species and community sequence in
the dataframe or matrix must be the same for all dataframe/matrices.
}
\details{
The function organizes the data for the functions of the package
\strong{SYNCSA}, placing the matrices of community, traits, phylogenetic
distance and environmental varibles in the same order.
}
\note{
The function organizes the data despite the absence of one of the
dataframes or matrices, provided that the community data had been entered. Unspecified
data will appear as NULL.
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\seealso{
\code{\link{matrix.t}}, \code{\link{matrix.x}},
\code{\link{matrix.p}}, \code{\link{syncsa}}
}
\keyword{SYNCSA}

