% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part.cor.R
\name{part.cor}
\alias{part.cor}
\title{First-order partial correlation coefficient}
\usage{
part.cor(rxy, rxz, ryz)
}
\description{
Internal function to obtain the first-order partial correlation coefficient.
}
\author{
Vanderlei Júlio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}

