% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut.row.matrix.R
\name{permut.row.matrix}
\alias{permut.row.matrix}
\title{Permutate rows in a matrix}
\usage{
permut.row.matrix(matrix, strata = NULL)
}
\description{
Internal function to permutate rows in a matrix.
}
\author{
Vanderlei Júlio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}

