% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ProgressBAR.R
\name{ProgressBAR}
\alias{ProgressBAR}
\title{Text Progress Bar}
\usage{
ProgressBAR(n, N, ...)
}
\arguments{
\item{n}{is a number}

\item{N}{is the total number}

\item{...}{additional arguments to txtProgressBar}
}
\description{
Text progress bar in the R console. See \code{\link{txtProgressBar}}.
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}

