% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SYNCSA-package.R
\docType{package}
\name{SYN_CSA}
\alias{SYN_CSA}
\alias{SYN_CSA-package}
\title{SYN_CSA}
\description{
This is the syncsa package!
}
\details{
SYNCSA is a package that implements methods developed by Pillar, Duarte and
others to measure functional trait convergence

measured with TCAP
}

