\encoding{UTF-8}
\name{SYNCSA-package}
\alias{SYNCSA-package}
\alias{SYNCSA}
\docType{package}
\title{
SYNCSA - Analysis of functional and phylogenetic patterns in metacommunities
}
\description{
Analysis of metacommunities based on functional traits and phylogeny of the community components. The functions that are offered here implement for the R environment methods that have been available in the SYNCSA application written in C++ (by Valério Pillar, available at http://ecoqua.ecologia.ufrgs.br/ecoqua/SYNCSA.html).
}
\details{
\tabular{ll}{
Package: \tab SYNCSA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-04-05\cr
License: \tab GPL-2\cr
}
}
\author{
Vanderlei Júlio Debastiani <vanderleidebastiani@yahoo.com.br>

Maintainer: Vanderlei Júlio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{package}
