% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.dummy.R
\encoding{UTF-8}
\name{var.dummy}
\alias{var.dummy}
\title{Generate dummy variable}
\usage{
var.dummy(data)
}
\arguments{
\item{data}{A data.frame.}
}
\value{
\item{data}{The data with all variables of class factor expanded.} \item{together}{A list
with sugestion to group of traits that are analysed together.}
}
\description{
Function to expand factor variables in dummy variables in a data.frame. See details.
}
\details{
Variables in the data.frame of class factor is expanded in dummy variables. Each
level of factor produce a new column in the dataframe, with presence (1) or
absense (0) of level. The name of columns is a combination of orginal variable name
plus the level separate by underscore ( _ ). Ordered factor and character class are
not expanded.
}
\seealso{
\code{\link{syncsa}}, \code{\link{organize.syncsa}}, \code{\link{var.type}}
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}
