% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var.type.R
\encoding{UTF-8}
\name{var.type}
\alias{var.type}
\title{Check the type of variables}
\usage{
var.type(data)
}
\arguments{
\item{data}{A data.frame or matrix.}
}
\value{
A vector with the variable types, where 'c' is continuous/numeric, 'o' is
ordinal, 'b' is binary, 'n' is nominal and 'f' is factor.
}
\description{
Function extracted (with small changes) of the function \code{\link{gowdis}}
to check the type of variables in a data.frame or matrix.
}
\seealso{
\code{\link{syncsa}}, \code{\link{organize.syncsa}}, \code{\link{var.dummy}}
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}
