% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartype.R
\encoding{UTF-8}
\name{vartype}
\alias{vartype}
\title{Check the type of variables}
\usage{
vartype(data)
}
\arguments{
\item{data}{A dataframe or matrix.}
}
\value{
\item{type}{The variable type.}
}
\description{
Function extracted (with small changes) of the function \code{\link{gowdis}} 
to check the type of variables in a dataframe or matrix.
}
\details{
The variable types, where 'C' is continuous/numeric, 'O' is ordinal, 'B' is
symmetric binary and 'N' is nominal.
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\seealso{
\code{\link{syncsa}}, \code{\link{organize.syncsa}}
}
\keyword{SYNCSA}

