\encoding{UTF-8}
\name{ADRS}
\alias{ADRS}
\docType{data}
\title{Artificial Data for Run SYNCSA}
\usage{
data(ADRS)
}
\description{Artificial data for run SYNCSA.}
\format{
  A list with 4 matrices.
  \describe{
    \item{commnunity}{Community data, observations of six species in ten sites.}
    \item{traits}{Matrix data of species described by two traits.}
    \item{phylo}{Matrix containing phylogenetic distance between species.}
    \item{envir}{Two environmental variables for each community.}
  }
}
\examples{
data(ADRS)
}
\keyword{datasets}