% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cent.norm.R
\encoding{UTF-8}
\name{cent.norm}
\alias{cent.norm}
\title{Matrix centralization and standardization}
\usage{
cent.norm(x, na.rm = FALSE)
}
\arguments{
\item{x}{A matrix}

\item{na.rm}{Logical argument (TRUE or FALSE) to specify if missing
observations are removed (Default na.rm = FALSE).}
}
\description{
Internal function for centralization and standardization in a matrix.
}
\seealso{
\code{\link{syncsa}}
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}
