% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cent.norm.R
\name{cent.norm}
\alias{cent.norm}
\title{Matrix centralization and standardization}
\usage{
cent.norm(x, na.rm = FALSE)
}
\description{
Internal function for centralization and standardization in a matrix.
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}

