\encoding{UTF-8}
\name{flona}
\alias{flona}
\docType{data}
\title{Hypothetical data for SYNCSA}
\usage{
data(flona)
}
\description{Hypothetical data for running examples.}
\format{
  A list with 4 matrices.
  \describe{
    \item{commnunity}{Community data, observations of 59 species in 39 sites.}
    \item{traits}{Matrix data of species described by five traits.}
    \item{phylo}{Matrix containing phylogenetic distance between species.}
    \item{environment}{Three environmental variables for each community.}
    }
}

\examples{
data(flona)
}
\keyword{datasets}
