% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part.cor.R
\encoding{UTF-8}
\name{part.cor}
\alias{part.cor}
\title{First-order partial correlation coefficient}
\usage{
part.cor(rxy, rxz, ryz)
}
\arguments{
\item{rxy}{Correlation between xy}

\item{rxz}{Correlation between xz}

\item{ryz}{Correlation between yz}
}
\value{
The first-order partial correlation coefficient.
}
\description{
Internal function to obtain the first-order partial correlation coefficient.
}
\seealso{
\code{\link{syncsa}}, \code{\link{cor.matrix}},
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}
