% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part.cor.R
\encoding{UTF-8}
\name{part.cor}
\alias{part.cor}
\title{First-order partial correlation coefficient}
\usage{
part.cor(rxy, rxz, ryz)
}
\arguments{
\item{rxy}{Correlation between xy}

\item{rxz}{Correlation between xz}

\item{ryz}{Correlation between yz}
}
\description{
Internal function to obtain the first-order partial correlation coefficient.
}
\author{
Vanderlei Júlio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}

