% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permut.row.matrix.R
\encoding{UTF-8}
\name{permut.row.matrix}
\alias{permut.row.matrix}
\title{Permutate rows in a matrix}
\usage{
permut.row.matrix(data, strata = NULL, seqpermutation = NULL)
}
\arguments{
\item{data}{A matrix.}

\item{strata}{Strata vector to specify restricting permutations in rows, must be the same length of number of rows in data matrix (Default strata = NULL).}

\item{seqpermutation}{A set of predefined permutation vector (Default seqpermutation = NULL).}
}
\value{
\item{permut.matrix}{The matrix permuted.} \item{samp}{The sequence of permutation.}
}
\description{
Internal function to permutate rows in a matrix.
}
\seealso{
\code{\link{syncsa}}, \code{\link{permut.vector}}
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}
