% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permut.row.matrix.R
\name{permut.row.matrix}
\alias{permut.row.matrix}
\title{Permutate rows in a matrix}
\usage{
permut.row.matrix(matrix)
}
\description{
Internal function to permutate rows in a matrix.
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}

