% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/permut.row.matrix.R
\encoding{UTF-8}
\name{permut.row.matrix}
\alias{permut.row.matrix}
\title{Permutate rows in a matrix}
\usage{
permut.row.matrix(matrix, strata = NULL)
}
\arguments{
\item{matrix}{A matrix}

\item{strata}{Argument to specify restricting permutations in rows
(Default strata = NULL).}
}
\description{
Internal function to permutate rows in a matrix.
}
\author{
Vanderlei Júlio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}

