% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressBAR.R
\encoding{UTF-8}
\name{ProgressBAR}
\alias{ProgressBAR}
\title{Text Progress Bar}
\usage{
ProgressBAR(n, N, ...)
}
\arguments{
\item{n}{Number of the current progress.}

\item{N}{Total number of cases.}

\item{...}{Other parameters for the txtProgressBar function.}
}
\description{
Text progress bar in the R console. See \code{\link{txtProgressBar}}.
}
\author{
Vanderlei Julio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}
