% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProgressBAR.R
\name{ProgressBAR}
\alias{ProgressBAR}
\title{Text Progress Bar}
\usage{
ProgressBAR(n, N, ...)
}
\description{
Text progress bar in the R console. See \code{\link{txtProgressBar}}.
}
\author{
Vanderlei Júlio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{SYNCSA}

