\encoding{UTF-8}
\name{SYNCSA-package}
\alias{SYNCSA-package}
\alias{SYNCSA}
\docType{package}
\title{
SYNCSA - Analysis of functional and phylogenetic patterns in metacommunities
}
\description{
Analysis of metacommunities based on functional traits and phylogeny of the community components. The functions, that are offered here, implement methods that have been available in the SYNCSA application written in C++ (by Valério Pillar, available at http://ecoqua.ecologia.ufrgs.br/ecoqua/SYNCSA.html) for the R environment.
}
\details{
\tabular{ll}{
Package: \tab SYNCSA\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab November 28, 2011\cr
Publication: \tab April 05, 2011 (Version 1.0)\cr
License: \tab GPL-2\cr
}
}
\author{
Vanderlei Júlio Debastiani <vanderleidebastiani@yahoo.com.br>

Maintainer: Vanderlei Júlio Debastiani <vanderleidebastiani@yahoo.com.br>
}
\keyword{package}
